Move: <VideoCard />, <VideoList /> 공통 컴포넌트 폴더로 이동 및 import 경로 수정
