+++ src/components/explore-page/index.ts
@@ -1,2 +1,1 @@
-export * from './video-list';
 export * from './pagination';
