+++ src/components/main-page/index.ts
@@ -1,3 +1,2 @@
 export * from './category';
-export * from './video-card';
 export * from './section-title';
