+++ src/components/shared/video-card/index.ts
