+++ src/components/shared/video-card/video-card.tsx
